[#451] Octo now performs type checks for `CASE` statements
Octo previously did not perform type validation for SQL `CASE` statements. This allowed for the creation of `CASE` branches with heterogenous types, causing correctness issues. These changes rectify this bug by adding type validation for `CASE` statements to the populate_data_type function.

This was done by adding type-checking if statements to the `cas_STATEMENT` and `cas_branch_STATEMENT` branches of the main switch in `populate_data_type`. These checks allow for SQL `NULL` types to be matched with any type, as this is the behavior expected by SQL clients, notably SquirrelSQL, as documented in various queries from test_squirrel_sql_connect_queries.bats.in.

Code duplication was minimized by the addition of a new `CAST_EQUIVALENT_TYPES` macro in populate_data_type.c. Also, a new `ERR_CASE_BRANCH_TYPE_MISMATCH` was added for reporting type errors in `CASE` statements.

New subtests were added to test_case_statement.bats.in to validate behavior for various edge cases, along with matching outrefs.

Note that in the case of nested `CASE` statements, the location of the syntax error is currently not reported due to the recursive nature of such cases. Specifically, syntax locations of child `CASE` statements are not known to parent statements, but the parent is where the syntax location/type mismatch error is issued.