Revised make install rules to put `_ydbocto.so` in `$ydb_dist/plugin/octo/o` instead of `$ydb_dist/plugin/o`
This included updates to `build.sh` and `test_helpers.bats.in` to account for new location of `_ydbocto.so`

Additionally, the TOZ004 test case was revised to expand paths in `$ydb_routines` by printing `$zroutines` from yottadb, allowing success when verifying against Octo initialization results. Indentation was also corrected for this test case.