Added EINTR handling to get_mem_usage .c and error handling for get_mem_usage function call in rocto.c
Previously, `get_mem_usagea did not handle an `EINTR` return value from `fgets` in `get_mem_usage.c`. This has been corrected so that this value of `errno` is caught and, in that case, the call to `fgets` retried.

Additionally, `rocto.c` did not handle error returns from `get_mem_usage`. This is now fixed by adding a check on the return value of this function and, if there is an error, issuing a new `ERR_MEMORY_USAGE` error message.