[#468] Resolved segfault for LIKE query comparing column references
As described in #468, Octo issues a SIG11 when comparing column references in SQL `LIKE` expressions.  This is now remedied by an addition `SqlValue` type check in `regex_specification.c` to ensure that `parse_literal_to_parameter` is only called on literals and not on column references.

This check is necessary since the aforementioned calls lead to SIG11s due to an unpopulated `parameter_index` field on `SqlValues` of type `COLUMN_REFERENCE` or, more generally, any non-literal `SqlValueType`.

Additionally, a new TRTE04 test case was added to confirm that this fix resolves the SIG11. However, this test fails a crosscheck with PostgreSQL due to an error. Accordingly, it has been temporarily revised to use an outref to validate that this expected error is returned until #419 is resolved.