[#402] Corrected failing Docker pipeline stage caused by recent Octo installation changes
Prior to resolving issue #402 (commit 537b45ed), the Octo Dockerfile relied on `src/aux/install.sh` to handle Octo installation when building the Octo Docker image. This file was subsequently removed in favor of using `make install` for Octo installations instead of an install script, per discussion on the aforementioned issue.

The deprecated `src/aux/install.sh` was removed from the Dockerfile, but this led to breakage because no revised installation method was put in its place. This led to a dangling `yes` command that would fill the container with `y`s and result in enormous artifact sizes and a failed `docker` stage in the pipeline.

To correct this, the Dockerfile was revised to use a new, pipeline-only `install.sh` for installing the requisite Octo files into the Docker image. This is necessary since `make install` has a dependency on CMake, which is not desirable for the Octo Docker image as it is not necessary after the build process and is not included in Ubuntu by default.

Moreover, the previous method of using the tarball produced by `make package` proved unviable due to the lack of native support for packaging shared library files using CPack from CMake. Since Octo requires `_ydbocto.so`, a tarball omitting this file will be insufficient for a functional installation.

To work within the above constraints as well as the limits of building Docker containers within the GitLab pipeline, the following high-level procedure was implemented:
1. Re-use the Octo build from preceding `ubuntu` jobs (run during the `build` pipeline stage) if they are `RelWithDebInfo` builds.
2. Use `install.sh` to copy the requisite files into the correct locations within the Docker image under construction
3. Perform Octo configuration in `tools/entrypoint.sh` whenever the Docker image run as a container

There are several items worth noting about the above procedure:
+ Not all files required for proper functioning of Octo are produced as part of the build process. There are some items that are done by `make install` that must be done manually in the absence of that command. Specifically, this means copying required, non-compiled files into the `build` directory artifact created by the initial `ubuntu` job. These include M routines from `src/aux`, `octo.conf` from `src/aux/octo.conf.default`, and the `postgres-seed.*` fixtures. These files are copied into the artifacts of `ubuntu` jobs by `tools/ci/build.sh` so that they can be made available to `tools/ci/install.sh` by the Dockerfile.
+ Similarly, the YDBPosix plugin is required by Octo, but is not included by default in the `yottadb-base` image. Accordingly, YDBPosix files have to be copied into the Docker image as well. These are also pulled from the previous `ubuntu` job in the same way as the files in the previous point.
+ `entrypoint.sh` was greatly expanded to ensure that Octo is properly configured before running `rocto` within a container spawned from the Octo Docker image. This entails creating and configuring a database with a separate OCTO region, as well as loading `postgres-seed.*` fixtures. Note that this script deletes and recreates the database to prevent issues when the container/image is run multiple times in succession, as operations like creating or updating regions must be run in order.
  - This means that we cannot blindly create a database whenever the container is spawned, as this will lead to a "database exists" error. On the other hand, attempting to update a database without the correct global directory setup leads to missing or redundant region errors. In short, the maintenance of state across image builds and container spawns means we must reset that state in order to ensure that running the container is idempotent to avoid these errors.

Due to the pipelined nature of these changes, it is not possible to test them automatically. Rather, the pipeline must be run to do so. These changes have been verified using a personal Docker account for the Docker Hub portions of the process. However, to confirm these changes work on master requires the credentials for the master Docker Hub account, and so such validation has been omitted for this MR.

During this process, it was discovered that it is possible to obscure operations done during the `docker build` process by setting certain command-line options when running the generated image with `docker run`. Specifically, mounting a volume with the `-v` option will supercede any filesystem contained in the image at the specified mountpoint. This will occur with no warning, which can lead to things silently breaking if the user isn't careful.

Specifically, `docker run` allows users to mount filesystems from their host machine to a Docker container at arbitrary locations. For example, the Octo README recommends adding ``-v `pwd`:/data`` to the `docker run` command, causing the current working directory to be mounted at `/data` in the container. This has the effect of making anything stored at `/data` inaccessible, without any indication that this has happened.

Consequently, if a user mounts a volume at a mountpoint that is used by the container, e.g. `/data`, it can lead to breakage when running the container. The fix for this is to either:
1. Not mount a filesytem to `/data`, or
2. Configure any mounted filesystem in a way that expected by the container. In this case, this means that `/data` must:
  + Have the file structure expected by YottaDB for a normal `$ydb_dir` setting, i.e. a structure like: `/data/$ydb_rel/g/`, etc.
  + Have a database properly configured for Octo usage, including enabling NULL_SUBSCRIPTS and ideally creating a separate region for Octo global variables. Details on this are available in the YDBOcto README.md.

In short, custom mounting directories from the host filesystem is possible with extra configuration, but is an expert friendly feature that is not required for normal users. Accordingly, these fixes are designed with the non-expert user in mind, so that a new YottaDB (and potentially new Docker) user can get things running with minimal effort.