[#481]  Reduced default verbosity settings to disable `INFO` messages unless `-v` is specified
Previously, Octo would issue an `ERR_LOADING_CONFIG` message at startup regardless of verbosity settings. However, since this message is called with the `INFO` macro, it should only appear for sufficiently high verbosity settings. This is corrected with this commit.

The reason for the erroneous behavior was that the `ERR_LOADING_CONFIG` message was issued prior to *any* verbosity level being set. Accordingly, any level set by the user would be ignored by this message. The fix was to revise the message into an `INFO_LOADED_CONFIG` message and place it *after* a verbosity level is set. These changes were to `octo_init.c`.

While making these changes, a bug was discovered in the application of command line verbosity settings in `parse_startup_flags.c`. Specifically, when using a number of `-v`s, the number of `v`s would be subtracted from the error severity level set at the time of the call to `parse_startup_flags`, rather than from a fixed value. This meant that the meaning of a certain number of `v`s would vary depending prior verbosity configuration, namely through `octo.conf`. The fix was to subtract the number of `v`s from the new default level of `ERROR` in all cases. This means that `-v` always means `WARNING` level, `-vv` always means `DEBUG`, and so on.

Note that these changes entailed decreasing the default verbosity level. Previously, `WARNING` was the default level set in `src/octo.conf.default`. However, this is more verbosity than is required for normal operation. Accordingly, the default verbosity level has been changed there to `ERROR` and in a few other relevant locations in the codebase. Related to this, `octo_init.c` was revised to disallow `FATAL` as a valid verbosity level, as users should not have the option to disable `ERROR` reporting. As part of this, the verbosity level checks in the `ERROR` and `FATAL` have been removed as they are now redundant.

Another significant change was the revision of the error messaging system to distinguish between error verbosity and severity levels. This entailed interface changes to `octo_log` and related macros as well as the division of the `ERROR_LEVEL` enum into two separate `VERBOSITY_LEVEL` and `SEVERITY_LEVEL` enums.

Additionally, the `ERR_BAD_CONFIG` message was revised to include the name of the config file being used. To support this revision, the `MERGE_CONFIG_PATH_AND_RETURN_ON_ERROR` macro was revised to populate the global `config->config_file_name` instead of using temporary/local variable. This change entailed updating a number of calls throughout `octo_init.c`.

Also, config file settings are now parsed in a new `parse_config_file_settings` function, which validates all loaded configuration files rather than the final merged version only. This results in configuration errors being flagged on a per-file basis whereas previously they were not. Similarly, an `INFO_LOADED_CONFIG` message is now issued for any configuration file loaded, not just the last one. Related to this, `octo_init.c` now stores settings from the initial call to `parse_startup_flags` and reuses them instead of calling this function twice. Also, zroutines processing is now done at the end of `parse_config_file_settings` to allow it to be skipped after the first, highest precedence value of zroutines is set.

To support the above changes, the order of configuration file parsing was reversed so that the highest precedence files are now parsed first, rather than last. This in turn required adding error handling logic for various libconfig calls that was previously missing so that the various failure states can be distinguished. To support this, a new `ERR_CONFIG_IO_FAILURE` error message and a new `CONFIG_ERROR_CHECK` macro were added. Also, `merge_config_file` was revised to distinguish between configuration files and the default configuration (represented as a string), which involved an interface change.

As part of the above change to parse order, `merge_config_file` and `OctoConfig` were revised to use a malloc'd pointer for maintaining the `config_file` struct. This is needed since we can no longer apply successively higher precedence settings to a constant base `config_t` by naively overwriting values as `merge_config_file_helper` does. However, that functionality can be maintained if we simply reverse the arguments to `merge_config_file_helper` so that the higher precedence settings overwrite the lower precedence ones such that only previously unpopulated settings are then set. The price for this is an additional `malloc` for the initial `config_t` member of `OctoConfig`, so that the `new_config_file` pointer created by each call to `merge_config_file` can be used to replace the previous value.

These changes also included the correction of a long-standing issue with configuration file syntax error handling. Specifically, such errors would produce messages that provided no helpful information whatever. This is now fixed by the following changes:
+ `BAD_CONFIG` error message format was revised for readability
+ A hard-coded value of `default` is now used for the error case of a call to `config_read_string`, as there is no filename to retrieve in this case. This change prevents the error message from printing `(null)` for the configuration filename.
+ Fixed unhelpful config parsing error message for failed `config_read_file` call by revising error information lookup calls to reference the correct configuration file structure
+ Added `TV008` test to verify above config syntax error handling fixes

Finally, a number of tests and outrefs were updated to reflect the verbosity changes. Tests revised included `test_multicmd_line.bats.in`, `test_octo_zroutines.bats.in`, `test_log_split.bats.in`, and `test_psql_connection.bats.in`. These were modified as the verbosity level previously used was actually relevant to the test. Similarly, `start_rocto` in `test_helpers.bash.in` was revised to use `-vvv` to prevent an infinite loop while waiting for a `rocto started` message. The outrefs updated include:
+ `TBP001.ref`
+ `TBP003.ref`
+ `TERR008.ref`
+ `TGB04.ref`
+ `TH003.ref`
+ `TH004.ref`
+ `TLS01.ref`
+ `TML002.ref`
+ `TML003.ref`
+ `TML004.ref`
+ `TML005.ref`
+ `TML006.ref`
+ `TML007.ref`
+ `TML009.ref`
+ `TML012.ref`
+ `TPGC002.ref`
+ `TSC18.ref`
+ `TSQ001.ref
+ `TSQ002.ref`
+ `TSQ003.ref`
+ `TSQ004.ref`
+ `TSQ005.ref`
+ `TSQ006.ref`
+ `TSQ007.ref`
+ `TSQ008.ref`
+ `TSQ009.ref`
+ `TSSCQ12.ref`
+ `TSSCQ13.ref`
+ `TSSCQ14.ref`
+ `TSSCQ16.ref`
+ `TSSCQ17.ref`
+ `TSSCQ18.ref`
+ `TSSCQ19.ref`
+ `TSSCQ21.ref`
+ `TSSLAC01.ref`

Also, several outrefs were updated to account for the temporary removal of `SSLRequest` messages from `rocto` log files (commit 519864dbac6d519df610c17f9fe4593901ffbcc9):
+ `TERR010.ref`
+ `TLS01.ref`
+ `TPC000.ref`

Additional miscellaneous changes:
+ Fixed capitalization of YDBPosix in install_posix.sh
+ Broke long `octo.default.conf` comments across multiple lines
+ Revised help message formatting in `parse_startup_flags.c` to be more readable
+ Revised `_ydboctoInit.m` to enforce verbosity for hardcoded `WARN` message
+ Renamed  `record_error_level` memeber of `OctoConfig` to `verbosity_level`