[#488] Added error handling for `ydb_init` calls so correct errors are emitted in case of failure
Previously, Octo would issue empty errors for various failure states emanating from from `ydb_init` in `octo_init.c`. This was due to the lack of error handling around that call, which is now added with this change.

As part of this change, `ydb_error_check` was revised to retrieve error messages via `ydb_zstatus` in the event of a `ydb_init` failure, as `ydb_data_s` cannot be used unless `ydb_init` succeedin the event of a `ydb_init` failure, as `ydb_data_s`/`ydb_get_s` cannot be used unless `ydb_init` succeeds.

To validate these changes, `TERR018` was added to `test_errors.bats.in`, along with a corresponding outref.