Corrected failing test by adding -f to rm command to prevent error if the target doesn't exist
A recent merge request (!596) submitted changes that included the addition of the TERR018 test case. This test case removes any `_ydbocto*.o` files to prevent interference with the various cases it tests. However, in some cases no files exist that match this pattern, causing the `rm` for their removal to fail and so fail the test. Adding `-f` to that `rm` call will prevent this error.

This change also includes an update to the TERR018 outref so that it accounts for recent YDB changes to the ICUSYMNOTFOUND message.