[#491] Added configuration option to require TLS connections from client
Previously, Rocto allowed clients to connect without TLS encryption if they did not request it. Now, Rocto has the ability to prohibit this behavior if desired, requiring clients to use TLS connections or else declining the connection.

To support this behavior, a new `ssl_required` configuration option was added to the list of options in `octo_init.c`, as well as to the defaults in `octo.conf.default`. Also, `rocto.c` was updated to check for this setting and issue a `FATAL` error and disconnect if `ssl_required` is set, but the client does not request a TLS connection.

Several tests were also added to `test_psql_connection.bats.in` to validate behavior of the new `ssl_required` setting. This also entailed adding a `notls` option to the `run_psql` helper to allow TLS to be selectively disabled even when it is available to the test system.

Additionally:
+ Some `INFO` messages that should only be logged locally were wrapped with the `LOG_LOCAL_ONLY` macro to make their lack of propagation explicit.
+ Updated Octo error documentation to more accurately reflect error meanings
+ Converted `FATAL` error in `parse_startup_flags` to `ERROR`
+ Converted `WARNING` messages to `TRACE` messages in `read_ssl_request.c`, including explanatory comments