Free config files after use
- Free config_file struct as well as the memory it points to
- Free config_file at end of rocto and octo main
- Free `new_config_file` if the file did not exist

Previously, rocto would call config_destroy on the config_file in merge_config_file, but it would not free the config_t allocation. The documentation states:

config_destroy() destroys the configuration config, deallocating all memory associated with the configuration, but does not attempt to deallocate the config_t structure itself.

Additionally, it would never free the last config_t parsed. This sets config->config_file properly so that it can be freed at the end of rocto's main.

Finally, if a config file did not exist, it would allocate memory for
the file, then returning without using that memory. This now
de-allocates that memory before returning.