[#372][#464] Fixed occasional failure in TPGC06
The fix for #372 and #464 (commit 7d898bec49be11918f09b59aee6c3a78d913aadc) included the introduction of `TPGC06`, which occassionally failed with output like the following:

```sh
> diff outref/TPGC006.ref clean_output.txt
4,5c4,5
< 0|0
< 5|5
---
> 0|6
> 5|56
```

The cause of this behavior turned out to be a logic error in the handling binary vs. text formats when creating `DataRow` messages in `send_result_rows.c` and `make_data_row.c`. Specifically, `send_result_rows` contained logic to populate the `DataRowParm` `length` member using either the length of the M value for the given row value (text format case) or the type size based on the `PSQL_TypeOid` for that row value.

However, this didn't match the expectations of a switch case in `make_data_row.c`, which used the length determined in `send_result_rows` when doing a `memcpy` from a `DataRowParm` value, which is always a string derived from an M value. In the case of a binary formatted integer (`PSQL_TypeOid_int4`), the length used for this `memcpy` would be `4`, even though the M value's length may *not* be `4`. This leads to erroneous output when if null terminator isn't fortuitously found at the correct offset.

Accordingly, the fix was to:
1. Move the logic for assigning format lengths from `send_result_rows.c` to the length determination loop at the start of `make_data_row`
2. Unconditionally set all `DataRowParm` `length` fields to the length of the M value they store in `send_result_rows.c`
3. Use `get_type_size_from_psql_type` in the `PSQL_TypeOid_int4` case in `make_data_row` to determine the size of the binary formatted value for the message, instead of expecting to use the `DataRowParm` `length` field.

Since this case relies on accidents of memory allocation/undefined behavior, it is not trivial to reproduce reliably and so no specific test case has been added to validate these changes. They have, however, been manually validated by running `TPGC06` 300 times with 0 failures. Prior to these changes, this test case would fail ~1-5 time in 100 runs.