[#511] Don't ignore config files passed with -c
Previously, Octo would ignore config files passed with -c:
it would assign the filename to a local variable, which would be lost
when returning from the current function. This change now assigns the
filename to a variable in `octo_init` so that it will be seen and read.

Additionally, this adds a bit more logging about which config files are
being read.

Closes https://gitlab.com/YottaDB/DBMS/YDBOcto/-/issues/511

- Print info message when loading config file specified with -c
- Don't set config_file_name to an uninitialized value if -c is not specified
- Add test for -c printing config info
- Give an error if an input file requested with -c is not found
- Add test for missing config file
- Don't special case config_file_name for logging