[#485] %ydboctoAdmin now prompts for a password twice when creating a new user
Previoulsy, `%ydboctoAdmin add user` would only prompt the user for a password once during user creation. This is not ideal, however, as a user may accidentally mistype the password the first time. Accordingly, a second password entry prompt was added to `%ydboctoAdmin` along with a check to confirm that the passwords entered into each prompt match before commiting the changes to the database.

As part of this change, the various `%ydboctoAdmin` BATS tests were updated to reflect this additional prompt, as was the `create_user` helper function in `test_helpers.bash.in`.

Also, a slight change was made to the documentation's wording explaining the use of the user creation functionality.