[#483] `CREATE TABLE` now allows for TAB and other non-printable characters in the `DELIM` qualifier
Previously, Octo would not accept non-printable characters such as TAB in the `DELIM` qualifier of a `CREATE TABLE` statement, since this would lead to a `LITNONGRAPH` error from YDB. This is now fixed by using `$CHAR` for such delimiter characters instead of literals.

More specifically, the syntax for `CREATE TABLE` statements was expanded to accept a list of ASCII integer values for use in the `DELIM` qualifier, e.g.:

```sql
CREATE TABLE DELIMNAMES (id INTEGER PRIMARY KEY, firstName VARCHAR(30), lastName VARCHAR(30), middleInitial VARCHAR(1), age INTEGER) DELIM (9, 9) GLOBAL "^delimnames(keys(""id""))";
```

This example specifies two TAB characters (ASCII value 9) to act as the internal delimiter of an Octo table. Note, however, that these delimiters are *not* applied to Octo output, which retains the default pipe `|` delimiter. The reason for this is that tables may be joined that have different delimiters, and so one common delimiter needs to be chosen anyway, so the default is used.

To implement this feature, a new rule was added to the `optional_keyword_element` rule to allow for specification of the aforementioned ASCII character list. Similarly, `delim_char_list` and `delim_char_list_tail` rules were added to support this new rule, as well as a new `SqlDelimiterCharacterList` struct in `octo_types.h` for representing this list in the parse tree.

Additionally, the `DELIM` rule stores an "is_dollar_flag" in the first byte of the delimiter string literal, using one of two new DELIM_IS_DOLLAR_CHAR and DELIM_IS_LITERAL macros. This value is later extracted in `tmpl_column_reference.ctemplate`.

In line with the above additions, `compress_statement.c` and `decompress_statement.c` were updated to account for the new `SqlDelimiterCharacterList` struct when storing and retrieving the binary parse tree for a given table.

Also, to account for this new syntax, `emit_create_table.c` was updated to include the delimiter character list in the generated `CREATE TABLE` string and, likewise, `hash_canonical_query.c` was expanded with a `delim_char_list_STATEMENT` rule.

To report any invalid ASCII characters passed to the `DELIM` qualifier, a new `ERR_INVALID_DELIM_CHAR` error message was added.

Finally, `tmpl_column_reference.ctemplate` was updated to extract a delimiter character list from the parse tree, if present, and pass it to a `$CHAR` call in the generated M plan.

A number of new tests were added to `test_createtable.bats.in`, `test_psql_connection.bats.in`, and `test_errors.bats.in` to validate the behavior of these changes in both Octo and Rocto, including the new possible error cases.

To implement the new rocto tests, the `start_rocto` and `run_psql` helper functions in `test_helpers.bash.in` were updated to take additional arguments. Specifically, `start_rocto` was corrected to actually accept an additional rocto command line argument, whereas previously this was not implemented correctly. `run_psql` was updated to accept SQL fixtures and feed them into `psql` to prevent repetition for doing so in individual test cases.

Note also that several of the new `test_errors.bats.in` outrefs are extra incremented to account for new test cases in the same test on other open merge requests in order to reduce conflicts.

Miscellaneous change:
    + Removed dead code in `handle_password_message.c`
    + Removed reliance on large buffer for storing delimiter in `create_table_defaults.c` and replaced `strncpy` with `MEMCPY_LIT`