[#442] Don't write one past the end of the input buffer
Previously, when writing a null terminator to the end of
`input_buffer_combined` in `readline_get_more`, if the line
filled up the entire buffer, the memory just past the end of the buffer
would be written to. This sets `cur_input_max` to one less than the
length of the input buffer so there is space to hold the null
terminator.