[#314] Add tests for sqllogic4
- Rewrite parsing to be reusable
- Add insert.py
- Add select4.test file
- Update outref
- Add ZWR header to sqllogic4.zwr
- Generate a unique id for the primary key

This avoids MAXNRSUBSCRIPTS for complicated joins by significantly
reducing the number of subscripts used.

  + Add unique index to the table for octo
  + Remove commented-out code
  + Add more comments
  + Make .zwr file reproducible with insert.py
  + Add GLOBAL to generated SQL statement

Note that sqllogic4.sql is not completely reproducible with insert.py,
it requires a few minor changes:

- Deleting CREATE INDEX statements
- Deleting DROP TABLE statements
- Deleting INSERT INTO statements

The reason for this is because the generated file is meant to be used
for both postgres-sqllogic4.sql and sqllogic4.sql, which have some
differences.

- Skip sqllogic tests with high memory usage

- Pad select filenames to 4 zeros

This makes the files sorted in order when using `ls`.

- Make sqllogic tests reusable

This significantly reduces the amount of code needed to add a new unit test
for sqllogic.

- Add TS prefix and link to exact file in sqllogic test name
- Greatly reduce the number of sqllogic tests run in a pipeline

Running all sqllogic4 tests takes over 10 minutes, slowing down
pipelines considerably and causing timeouts. Instead this only runs a
random sample of 3% of the tests.