[#442] Address clang compiler warning by converting `else if` check into an `else` plus `assert`
Previously, `handle_bind.c` distinguished between `Bind` and literal parameters using an `if`/`else if` check in its main loop. However, this created the implicit possibility of an uninitialized variable access due to the lack of an `else` clause. This is now remedied by converting the `else if` to an `else` clause, and moving the relevant check from the `else if` into an `assert` in the new `else` block.

To test this change, `TPGC004` was expanded to include a single parameter/single `Bind` parameter case.

Additionally, comments were added/expanded in `handle_parse.c` and `handle_bind.c` to further clarify the `Bind` parameter population process.

As a miscellaneous change, `handle_parse.c` was significantly revised to avoid most `malloc`/`free`s in favor of stack allocations and to eliminate needless string copies.