[#442] Give an error if the startup message does not have a username
- Add a new ERR_ROCTO_MISSING_USERNAME which returns PSQL_Code_Connection_Exception to the client
- Check if the startup message has no parameters
- Check if there is at least one parameter, but no username
- Add missing_username test to test_handle_password_message

This also re-enables the `test_handle_password_message` test,
since it was commented-out before. Additionally it makes it possible to
create a dummy startup message without a username.

- Fix message test to match protocol

If no parameters are given, it is an error, since the username is always
required in the startup message. The test incorrectly expected this to
succeed.

- Remove unnecessary __wrap_md5_to_hex

`md5_to_hex` works provided the database access it relies on in
handle_password_message is correctly mocked. Since we already require
openssl for Octo, and there's no possibility that md5_to_hex could fail other than memory exhaustion, there is no need to mock it on each call.
This makes the test code much shorter to write.

- Use a macro instead of a constant variable

This allows separating declaration from initialization by using a
constant array instead of a VLA.

- Remove unnecessary initialization and separate initialization from declaration