[#442] Add clang-tidy to build pipeline
- Add clang-tidy to gitlab environment

Note that the gitlab environment does _not_ use the Dockerfiles; this
adds it directly in `.gitlab.yml`.

- Print the warnings that differ
- Add clang as well as clang-tidy

Without this, there will be lots of spurious `stdarg.h` errors.

- Update centos ref to not include `note:`
- Don't print 'x warnings generated' while running clang-tidy

This removes ~100 lines of unhelpful messages.

- Use different .ref files for debug and release

In release mode, `assert`s are compiled out and clang-tidy will emit false positives.

- Don't run clang-tidy on CentOS

CentOS 7 does not have clang-tidy available, and is the version used so
that YDB can test backwards compatibility.