[#320] Treat empty column values as SQL `NULL`
Previously, Octo ignored `NOT NULL` constraints in `CREATE TABLE` statements and treated all empty column values as empty strings. However, this is incorrect in many cases since an empty string is considered a positive value in SQL, whereas `NULL` is the absence of a value. Accordingly, this commit enhances Octo to treat empty fields as SQL `NULL`, unless `NOT NULL` and/or `PRIMARY KEY` is specified in the DDL statement. In that case, empty string values are converted to a positive "empty" value based on the column's type per the following scheme:

1. `VARCHAR`s remain the empty string
2. `NUMERIC`s and `INTEGER`s are converted to 0
3. `BOOLEAN`s are converted to `false`, i.e. 0

To perform this conversion, a new `empty2null` helper function was added to `aux/_ydboctoplanhelpers.m`. This function conditionally converts empty string values to `$ZYSQLNULL` or, if `NOT NULL` was specified in the DDL, to 0 or the empty string per the scheme above.

This conversion is performed in `tmpl_column_reference.ctemplate` by wrapping calls to `$PIECE` that lookup column values. The conditions described above are pulled from the DDL via a `get_keyword` call to lookup `NOT_NULL` (or `PRIMARY_KEY`) status. Similarly, this `NULL` conversion should not be performed on cross-references as these are predicated on literals, which are retained in this case. Cross-reference status for a column is also determined via a `get_keyword` call on `OPTIONAL_XREF_INDEX`.

Support for `NULL` columns was also added to `rocto` by revising `make_data_row` to use a column data length of `-1` for such columns per the entry for `DataRow` messages in the [PostgreSQL documentation](https://www.postgresql.org/docs/11/protocol-message-formats.html).

Several test cases were added to `test_createtable.bats.in` to validate the new changes, along with `TJC012` in `test_jdbc_connection.bats.in` to validate a previously failing query from the query generator. Also, `simple_go_select_all` and `TPGC05` were updated to account for the fact that the `pq` Go driver cannot handle `NULL` columns, while `TWO06` was updated to use a fixture instead of a hardcoded query. Further, a number of outrefs were updated to account for this change, many of which were due to changes in generated plans, which now include the `empty2null` call:
+ `TCF018`
+ `TCK03`
+ `TCT011`
+ `TEO03`
+ `TGB04`
+ `TJC010`
+ `TJC011`
+ `TNK01_errors`
+ `TOB04`
+ `TOB05`
+ `TOB09`
+ `TOB11`
+ `TPGC000`
+ `TPGC002`
+ `TPGC005`
+ `TSC19`
+ `TSSCQ10`
+ `TSSCQ11`
+ `TSSCQ12`
+ `TSSCQ13`
+ `TSSCQ16`
+ `TSSCQ17`
+ `TSSCQ18`
+ `TSSCQ19`
+ `TSSCQ20`
+ `TW06`
+ `TWI01`
+ `TWI02`
+ `TWI03`
+ `TWI04`
+ `TWI05`
+ `TWI06`
+ `TWI07`
+ `TWI09`
+ `TWO01`
+ `TWO02`
+ `TWO03`
+ `TWO04`
+ `TWO05`
+ `TWO06`
+ `TWO07`
+ `TWO08`
+ `TWO09`
+ `TWO10`
+ `TWO11`
+ `TWO12`
+ `TWO13`
+ `TWO14`
+ `TWO15`

Additionally, some Octo/PostgreSQL crosscheck test cases began failing due to differences between the output formatting between `psql` and the JDBC driver. Specifically, the JDBC driver populates `NULL` column values with the string literal `null` whereas `psql` simply leaves these columns blank. To work around this issue, `run_query_in_octo_and_postgres_crosscheck` was revised to use the JDBC driver for both Octo and PostgreSQL queries when specified, instead of only using the JDBC driver for the Octo queries.

However, the above change introduced some PostgreSQL authentication issues in the pipeline, as `psql` was not configured for usage with the JDBC driver. To resolve these issues, the `create_postgres_database` helper function was revised to grant database permissions to `$USER` on the creation of each database and the `test_jdbc_connection.bats.in` file was updated to create an Octo user for `$USER` as well for this case. Similarly, various Java fixtures were revised to use the system user id instead of a hardcoded `ydb` user.

Also as part of these authentication changes, a manual setting of `USER=root` in the `init_tls` helper function was removed as it needlessly multiplied the number of user ids used in the test system. Further, `build.sh` was updated to force PostgreSQL to use password authentication instead of `Ident` authentication by updating `pg_hba.conf`.

Additionally:
+ Various cosmetic changes
+ `test_inner_join_order_by.bats.in` was updated to include test numbers for each case to make targeted runs easier
+ Added test numbers to `test_psql_authentication.bats.in`
+ A new `ERR_ROCTO_USER_LOOKUP` error message was added to output user information when attempts to lookup user credentials fails
+ `test_psql_authentication.bats.in` was updated to account for the above new error message
+ Added user creation to `test_squirrel_sql_connection_queries.bats`
+ Revised `TMU02` to create Octo user from $USER and abstract the now variable username from the outref
+ Updated Northwind database to use `NULL` value
+ Changed `GVUNDEF`/`LVUNDEF` check in `handle_password_message.c` to `GVUNDEF` only as the relevant `ydb_get_s` is always for a GVN. Updated `test_handle_password_message` accordingly
+ Revised `hello_psql.bats.in` and `load_postgres_fixture` helper function to store command output for reference when debugging
+ Corrected error due to use of backticks in comment in `create_postgres_database`