Reduce test_jdbc_connection test runtime to around 25% by running fewer queries
Previously, the `test_jdbc_connection.bats.in` tests would take several minutes to complete running in the pipeline due to a large number of random queries being used in several subtests. To reduce this runtime, the `run_query_generator` helper function was revised to run only 1/4 of the queries normally run by the query generator.