[#536] If clang-format doesn't exist in the environment, try a version suffix
This fixes pipeline errors like the following:

```
./tools/ci/build.sh: line 86: clang-format: command not found
```

- Install clang-format-9 on Ubuntu
- Try to discover if a specific version is available
- Use the version we just discovered
- Fail the pipeline if not on CentOS 7 and a recent clang-format was not found
- Search for clang-format in pre-commit hook as well as on CI

This prevents the same 'clang-format not found' errors as seen in the
pipeline.

- Allow missing clang-format on RHEL, not just CentOS
- Try to find higher versions of clang-format before looking for lower
versions

- Don't try to run clang-format if it doesn't exist
- Format files automatically in pre-commit hook
- Don't use a second grep for diff-files

`git diff-files` by default always return 0 even if a match was found.
It has an `--exit-code` option which allows telling if there were
changes; this avoids another needless command.

- Don't check files unless they are staged to be committed

This avoids giving errors that are not related to the current commit.
Note that this does not handle the case when only part of a file is
staged to be committed.